Factor out express middlewares
