+++ package.json
@@ -7,10 +7,7 @@
     "type": "git",
     "url": "git+https://github.com/vacuumlabs/cardano.git"
   },
-  "keywords": [
-    "cardano",
-    "wallet"
-  ],
+  "keywords": ["cardano", "wallet"],
   "author": "Rafael Korbas",
   "license": "ISC",
   "bugs": {
@@ -18,7 +15,8 @@
   },
   "scripts": {
     "start-server": "node server/index.js",
-    "dev": "PORT=${PORT:-3000} node server/index.js & webpack --watch --config webpack.build.config.js",
+    "dev":
+      "PORT=${PORT:-3000} node server/index.js & webpack --watch --config webpack.build.config.js",
     "eslint": "eslint .",
     "build": "webpack --config webpack.build.config.js --progress",
     "test": "webpack --config webpack.test.config.js",
