+++ server/index.js
@@ -1,58 +1,22 @@
 const cors = require('cors')
 const bodyParser = require('body-parser')
 const express = require('express')
-const redis = require('redis')
+const statsMiddleware = require('./middleware/stats')
+const forceHttpsMiddleware = require('./middleware/forceHttps')
 const config = require('./helpers/backendConfigLoader')
 const app = express()
-const client = redis.createClient(config.REDIS_URL)
-
-const knownIps = new Set()
-
-const monthNames = [
-  'January',
-  'February',
-  'March',
-  'April',
-  'May',
-  'June',
-  'July',
-  'August',
-  'September',
-  'October',
-  'November',
-  'December',
-]
 
 app.use(bodyParser.json())
 app.use(cors())
 
 // must be before every other route to guarantee the redirect!
 if (config.CARDANOLITE_FORCE_HTTPS === 'true') {
-  app.use((req, res, next) => {
-    if (!req.secure && req.get('x-forwarded-proto') !== 'https') {
-      res.redirect(301, `https://${req.get('host')}${req.url}`)
-    } else {
-      res.setHeader('Strict-Transport-Security', 'max-age=31536000')
-      next()
-    }
-  })
+  app.use(forceHttpsMiddleware)
 }
+app.use(statsMiddleware)
 
 app.use(express.static('public_wallet_app'))
 app.use(express.static('public_landing_page'))
-// stats middleware
-app.use((req, res, next) => {
-  // 'x-forwarded-for' due to internal heroku routing
-  const ip = req.headers['x-forwarded-for']
-  if (!knownIps.has(ip)) {
-    const date = new Date()
-    knownIps.add(ip)
-    client.incr('total')
-    client.incr(`${monthNames[date.getMonth()]}-${date.getFullYear()}`)
-    client.incr(`${date.getDate()}.${date.getMonth() + 1}.${date.getFullYear()}`)
-  }
-  next()
-})
 
 require('./blockchainExplorerProxy')(app)
 require('./transactionSubmitter')(app)
